+++ css/style.css
@@ -1369,7 +1369,7 @@ main {
 
 @media (min-width: 1024px) {
   .img-upload__wrapper {
-    padding-bottom: 0;
+    padding-bottom: 60px;
   }
 }
