+++ js/main.js
@@ -1,6 +1,7 @@
+import './img-upload.js';
+import './img-upload-validation.js';
 import {createDataForGallery} from './data.js';
 import {formGalleryAsFragment} from './gallery.js';
-import './img-upload.js';
 
 document.querySelector('.pictures').append(
   formGalleryAsFragment(createDataForGallery(25))
